src/common/requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
